refactor: BookService를 application 패키지로 이동
